Se usa un solo modal para eliminar las notas
